Refactor, cleaning up and splitting into different files, by codex
