+++ src/App/frontend/src/features/validation/callbacks/onFormSubmitValidation.ts
@@ -3,8 +3,8 @@ import { useCallback } from 'react';
 import { ValidationMask } from '..';
 
 import { FormStore } from 'src/features/form/FormContext';
-import { useGetNodesWithErrors } from 'src/features/validation/derivedValidations';
 import { useWaitForValidation } from 'src/features/validation/validationContext';
+import { useGetNodesWithErrors } from 'src/features/validation/validationHooks';
 import { useOurEffectEvent } from 'src/hooks/useOurEffectEvent';
 
 /**
