+++ src/App/frontend/src/features/validation/callbacks/onPageNavigationValidation.ts
@@ -2,9 +2,9 @@ import { useCallback } from 'react';
 
 import { useRefetchInitialValidations } from 'src/core/queries/backendValidation';
 import { FormStore } from 'src/features/form/FormContext';
-import { useValidationsSelector } from 'src/features/validation/derivedValidations';
 import { getVisibilityMask } from 'src/features/validation/utils';
 import { useWaitForValidation } from 'src/features/validation/validationContext';
+import { useValidationsSelector } from 'src/features/validation/validationHooks';
 import { usePageOrder } from 'src/hooks/useNavigatePage';
 import { useOurEffectEvent } from 'src/hooks/useOurEffectEvent';
 import type { PageValidation } from 'src/layout/common.generated';
