+++ src/App/frontend/src/features/validation/deriveNodes.test.ts
@@ -1,0 +1,59 @@
+import { defaultDataTypeMock } from 'src/__mocks__/getUiConfigMock';
+import { processLayouts } from 'src/features/form/layout/LayoutsContext';
+import { makeLayoutLookups } from 'src/features/form/layout/makeLayoutLookups';
+import { MissingRowIdException } from 'src/features/formData/MissingRowIdException';
+import { deriveNodes } from 'src/features/validation/deriveNodes';
+import type { FormStoreState } from 'src/features/form/FormContext';
+import type { ILayoutCollection } from 'src/layout/layout';
+import type { ExpressionDataSources } from 'src/utils/layout/useExpressionDataSources';
+
+describe('deriveNodes', () => {
+  it('throws when a repeating row is missing its row ID', () => {
+    const layoutCollection = {
+      Form: {
+        data: {
+          layout: [
+            {
+              id: 'group',
+              type: 'RepeatingGroup',
+              children: ['input'],
+              dataModelBindings: {
+                group: { dataType: defaultDataTypeMock, field: 'Group' },
+              },
+            },
+            {
+              id: 'input',
+              type: 'Input',
+              dataModelBindings: {
+                simpleBinding: { dataType: defaultDataTypeMock, field: 'Group.Value' },
+              },
+            },
+          ],
+        },
+      },
+    } satisfies ILayoutCollection;
+    const layouts = processLayouts(layoutCollection, defaultDataTypeMock);
+    const state = {
+      bootstrap: {
+        layoutLookups: makeLayoutLookups(layouts, layoutCollection),
+      },
+      data: {
+        models: {
+          [defaultDataTypeMock]: {
+            debouncedCurrentData: {
+              Group: [{}],
+            },
+          },
+        },
+      },
+    } as unknown as FormStoreState;
+
+    expect(() =>
+      deriveNodes(state, {
+        pageOrder: ['Form'],
+        pdfLayoutName: undefined,
+        hiddenDataSources: {} as ExpressionDataSources,
+      }),
+    ).toThrow(new MissingRowIdException('Group[0]'));
+  });
+});
