+++ src/App/frontend/src/features/validation/selectors/componentValidationsForNode.ts
@@ -1,7 +1,7 @@
 import { useMemo } from 'react';
 
 import { FormStore } from 'src/features/form/FormContext';
-import { useVisibleValidations } from 'src/features/validation/derivedValidations';
+import { useVisibleValidations } from 'src/features/validation/validationHooks';
 import { useIndexedId } from 'src/utils/layout/DataModelLocation';
 import type { ComponentValidation, NodeRefValidation } from 'src/features/validation/index';
